% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.bricks.R
\docType{package}
\name{simmer.bricks-package}
\alias{simmer.bricks-package}
\title{\pkg{simmer.bricks}: Helper Methods for \pkg{simmer} Trajectories}
\description{
Provides wrappers for common activity patterns in \pkg{simmer} trajectories.
}
\seealso{
\pkg{simmer}'s homepage \url{https://r-simmer.org} and
GitHub repository \url{https://github.com/r-simmer/simmer.bricks}.
}
\author{
Iñaki Ucar
}
