% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_until.R
\name{wait_until}
\alias{wait_until}
\title{Wait until a Signal is Received}
\usage{
wait_until(.trj, signals)
}
\arguments{
\item{.trj}{the trajectory object.}

\item{signals}{signal or list of signals, accepts either a string, a list of strings or a
callable object (a function) which must return a string or a list of strings.}
}
\value{
Returns the following chain of activities: \code{\link[simmer:send]{trap}}
> \code{\link[simmer:send]{wait}} > \code{\link[simmer:send]{untrap}}
(see examples below).
}
\description{
This brick encapsulates a stop: trap one or more signals, wait for them,
then untrap the signals and continue the trajectory.
}
\examples{
## These are equivalent:
trajectory() \%>\%
  wait_until("green")

trajectory() \%>\%
  trap("green") \%>\%
  wait() \%>\%
  untrap("green")

}
